facial_recognition/requirements.txt
Flask==2.0.2
flask-cors==4.0.1
face-recognition==1.3.0
numpy==1.22.3
requests==2.26.0
python-dotenv==0.19.1
